requirements.txt
tensorflow==2.12.0
numpy==1.23.5
matplotlib==3.7.1
Pillow==9.5.0
